Fix vector type for DISPLAY_OFFSET

+++ bedrock/bedrock-codec/src/main/java/org/cloudburstmc/protocol/bedrock/data/entity/EntityDataTypes.java
@@ -63,7 +63,7 @@ public class EntityDataTypes {
     };
     public static final EntityDataType<BlockDefinition> DISPLAY_BLOCK_STATE = new EntityDataType<BlockDefinition>() {
     };
-    public static final EntityDataType<Vector3f> DISPLAY_OFFSET = new EntityDataType<Vector3f>() {
+    public static final EntityDataType<Vector3i> DISPLAY_OFFSET = new EntityDataType<Vector3i>() {
     };
     public static final EntityDataType<Byte> CUSTOM_DISPLAY = new EntityDataType<Byte>() {
     };
